topography: changed computation of hazard
Previously, there was an error made. The hazard has to be max reduced
only within a single flow and then it has to be added to the total hazard.
Before, it was the max over all flows touching a cell, which is wrong.

In topography the `compute_hazard` was changed to `compute_hazard_flow`
which only computes the hazard within a single flow now. Then the
addition is handeled within `Simulation::run`.