Simulation: refactored file saving in simulation
Simulation now has a get_file_handle function that returns a unique_ptr
to the topography_file base class. This should now be used to construct
the output file handles.

Depending on input settings the file handle will either be an AscFile
or a NetCDFFile under the hood.

The File IO settings are controlled by a struct in the input parameters.
When using the `save` method, the suffix gets appended automatically and
only the base name should be entered.
This helps to distinguish between the .asc and .nc file endings.
Unfortunately, this currently breaks the names of the masked files,
because they already contain a period.

Also the Output enum in topography_file has been renamed to
OutputQuantity.

Co-authored-by: Amrita Goswami <amrita16thaug646@gmail.com>