Volume correction: Added implementation
A volume correction is applied every time a lobe is added. The volume
deficit, (compared to the true volume) is applied to the intersecting
lobes, weighted by their area fractions, in Topography::add_lobe.
By default, this is true. However, this can be set via an input
parameter called `volume_correction`.